Updated pgMonitor dependencies and compatibility.
This brings the container suite inline with the requirements for
pgMonitor v3.2, with a few differences noted below.

The following versions are now utilized:

- pgMonitor 3.2
- Prometheus 2.9.2
- Grafana 6.3.4

The postgres_exporter is still pegged at 0.4.7.

This commit does not utilize the OS, etcd, or Prometheus dashboards
that pgMonitor makes available as we do not ship the node_exporter
(which would not provide accurate metrics as to how a PostgreSQL
container is affecting a system) or prefer to let the user decide
how they wish to monitor additional services.