Updates ports on the collect and pgbadger containers.
The service ports for pgbadger and postgres-exporter are currently
hardcoded in the defauts for each. This commit updates the images
and start scripts so that the port can be set by an envrionment
variable. The examples are also updated to set the EXPORTER_PORT
in the container environment on creation.