Updated the 'crunchy-postgres', 'crunchy-collect' and 'crunchy-grafana' containers as needed to support Grafana v6.3.4 and pgMonitor v3.2.
For the 'crunchy-grafana' container, this includes
creating the 'notifiers' directory during container
initialization to avoid a missing 'notifiers' directory
error during Grafana startup.

For the 'crunchy-collect' container, this includes
ensuring the PGBACKREST_INFO_THROTTLE_MINUTES
env var is properly set in the PostgreSQL Exporter
queries file.

For the 'crunchy-postgres' container, this includes
creating and leveraging a custom shell script that
can be utilized to obtain pgBackRest info, and
ensuring the setup SQL file run during initialization
of the pgMonitor module is properly configured to
use the custom script.

Additionally, the metrics example has been updated
to remove port 9100 from the crunchy-collect
containers.