Allow for customizing timeouts around specific pg_ctl commands.
This introduces three additional parameters for managing timeouts around
three specific `pg_ctl` commands:

- PG_CTL_START_TIMEOUT
- PG_CTL_STOP_TIMEOUT
- PG_CTL_PROMOTE_TIMEOUT

which coorespond with the "start", "stop", and "promote" pg_ctl commands
respectively.

These interface with the PostgreSQL "PGCTLTIMEOUT" environmental variable
and override the PostgreSQL default value as needed.

This is particularly helpful for "stop" events in a Kubernetes environment
to ensure that a pod does not terminate before PostgreSQL has fully flushed
its data to disk.

Future work in Kubernetes environments should better utilize the
`terminationGracePeriodSeconds` pod variable so that it can automatically
set the timeout time based on this value.