Explicitly set the defaults for the PG_CTL_*_TIMEOUT variables.
Testing within the containerized runtimes showed that PostgreSQL will treat
a "PGCTLTIMEOUT=" as something that is 0-like, and would cause the container
to not start.

(More interestingly, outside of the container, even with the error message,
PostgreSQL would boot up shortly thereafter).

The easiest/safest option for the time being was to default to the PostgreSQL
default, which is 60 seconds.