Ensure crunchy-collect / pgMonitor facilities start up on PG 9.5/9.6
3896128a1, which updated the Crunchy Container Suite to use
pgMonitor v3.2, added in additional version checking logic to load
the correct set of queries that pgMonitor relies on to scrape metrics.
A bug was introduced in this logic where any supported version of
PostgreSQL below version 10 would be treated as an unsupported version.

This fix adds in the appropriate version strings to check for 9.5/9.6