Makefile and Dockerfiles for all Postgres HA and Postgres GIS HA Images
Currently, the Crunchy Container Dockerfiles are broken down by the image
operator system (Centos or RHEL) and then by Postgres version. To reduce
redundancy and simplify testing, this can be reduced to just different
files for each OS.

Updates to the existing Makefile for the postgres-ha and postgres-gis-ha
targets, as well adding runtime variable defaults for use when building
the mentioned images. Makefile also dynamically sets the appropriate
string for use when installing pgAudit.

New Dockerfiles for Centos and RHEL versions of both the postgres-ha
and postgres-gis-ha containers allow for image creation with any supported
Postgres version (currently 9.5, 9.6, 10 and 11). Where the PG version
would have a period (e.g. 9.5 or 9.6) and an associated library or
repository does not (for instance, the pgdg96 repo), the version period
is removed during the Dockerfile's associated RUN sequence.

Issue: [ch6188]