Implements the workflow for converting an existing database to a Patroni cluster per the steps defined in the Patroni documentation.  This includes first manually starting the database with 'pg_ctl start', creating required Patroni credentials (i.e. superuser and replication credentials), and then using Patroni to restart the database to update and configuration as required.
The container now also listens for the initialization of the
Patroni cluster in order to do certain post-initialization
tasks.  This includes the initilization of pgBackRest, as
well as the creation of the 'crunchyadm' user that will
be utilized by the 'crunchy-admin' sidecar.

A new 'crunchy-admin' sidecar container has also been
created.  This container is intended to share a volume
mount with with a primary or replica container, containing
a unix socket that can be utilized for accessing the
database.  Specifically, the 'crunnchy-admin' container
can utilize this socket to do local peer authentication
into the PG database.

Also, pgBackRest is now utilized for all replica creation
(a fallback to pg_basebackup is still included to allow
replica creation in emergency situation in which
pgBackRest is not functioning).  To support this, pgBackRest
initialization now also includes the automatic creation
of the initial full pgBackRest database backup for the
cluster.  This backup can then be utilized by replicas,
specifically when performing a pgBackRest delta restore
during initialization of the cluster.

Also, please note that the current 'postgres-ha' example
may not work with persistent volumes that are
'readWriteOnce', depending on cluster architecture
and configuration.  This is because both the primary
and replicas require access to the same PVC
containing the pgBackRest repo, i.e. the primary
to , and with 'readWriteOnce' volumes multiple
nodes cannot mount the same volume.  Therefore,
if the primary and replicas are scheduled on
seperate nodes, they may not be able to mount
the required volumes.  This will be addressed once
the pgBackRest dedicated repo container is moved
from PGO to the Container Suite, and can therefore
be included in the example.