Remove the "PITR" functionality from "crunchy-pgbasebackup-restore".
This functionality would work if `ARCHIVE_MODE` is enable on the
"crunchy-postgres" and the archives are being stored to a volume that is
accessible by the "crunchy-postgres" container that is being restored. However,
the `ARCHIVE_MODE` functionality uses a recovery command that has known issues
in production environments. This will be fully replaced by the archiving
functionality in pgBackRest.

As part of the changes to archiving, and in part to make the
"crunchy-postgres" family of containers compatible with PostgreSQL 12, this
commit removes the flags that would allow the "crunchy-pgbasebackup-restore"
container set up a "crunchy-postgres" container to perform a
point-in-time-recovery. These include:

- `RECOVERY_TARGET_NAME`
- `RECOVERY_TARGET_TIME`
- `RECOVERY_TARGET_XID`
- `RECOVERY_REPLAY_ALL_WAL`

If you need to enable point-in-time-recovery for your application, please use
the pgBackRest integration with the containers.

This commit does not remove `ARCHIVE_MODE` as the main goal is to move the
containers project to have PostgreSQL 12 compatibility. Changes to
`ARCHIVE_MODE` will occur in a future commit.

Issue: [ch6213]