Update Postgres GIS HA example
This commit updates the Postgres GIS HA example to match the
latest updates from the Postgres HA example. The main differences
are in labeling the various resources as 'postgres-gis-ha' in place
of 'postgres-ha' as well as updating the PATRONI_SCOPE environment
variable to use 'example-gis-cluster' in place of 'example-cluster'
so that both the Postgres HA and Postgres GIS HA examples can be
run simultaneously.