update prometheus kubernetes configuration
- set 'instance' label to kubernetes pod label 'deployment-name'.
- set 'job' label to kubernetes pod label 'pg-cluster'.

These prometheus labels were not helpful in query filtering as they did
not provide meaningful or reliable values. In particular, the regex for
'instance' would not allow for cluster names that included a '-'. As
well, 'job' should be at the 'cluster' level.

The purpose of these changes are to improve query filtering abilities as
well as to achieve the goal of the 'job' and 'instance' labels as
described here:

https://prometheus.io/docs/concepts/jobs_instances/

Specifically, that a 'job' is a collection of instances with the same
purpose, in the same way that a 'cluster' is a collection of identical
postgres instances.