Add support for PostgreSQL 12.
This adds support for PostgreSQL 12 across all supported container platforms.

The biggest change in this release is the support for how PostgreSQL 12
handles standby/recovery mode. The "recovery.conf" file is no longer supported
by PostgreSQL. Instead, a user puts their standby/recovery settings into the
main "postgresql.conf" file, and indicates to the PostgreSQL instance what
mode to use based on the presence of a "standby.signal" or "recovery.signal"
file.

This accounts for this change, as well as enables the container builds across
the supported platforms.

Issue: [ch6213]