updates versions for latest PostgreSQL release
The version numbers were pulled from a commit that existed prior
to the most recent release of PostgreSQL.

Additionally, the order of pgimages was updated to encourage
building images which populate cache ahead of others.
This change is only harmful to those who seek alpha order and
may benefit -jN execution.