Add Automated Tests for Postgres HA and Postgres GIS HA Examples
In keeping with the existing integration test suite, new tests are
added for the Postgres HA and Postgres GIS HA container examples.
These tests run the previously developed examples that deploy two
pods with the Patroni HA solution configured. The tests will run as
follows:

-In a new namespace, check that the relevant deployments and pods are
  created and ready
-Test the database connectivity on the first pod
(-For the GIS example, verify that the postgis extension is installed)
-Return the Patroni status information and validate the pods show as
  running and there is exactly one leader assigned to the Patroni cluster
-Cleanup the created resources

This ensures that the basic operation of the relevant examples are
working as expected.