Revert "An kubectl wrapper script is now provided that"
This reverts commit 90072abe8c719b17364cbe892f726e7ac5e04846.

This effectively removes the kube-ssh-wrapper.sh wrapper script
utilized as a drop-in replacement for SSH for use by pgBackRest, as
well as any bootstrap commands run in support of this solution (i.e.
the creation of the 'pod_env.sh' file that is utilized by the script).
This also includes removing the install of 'kubectl' from all Dockerfiles
since it will no longer be needed with SSH re-enabled.