removes unnecessary docker.io tagging of images
As we were unable to identify the source of this build step
nor its impact on the release builds, we are removing this step.

Further, it prevents using the PGO_IMAGE_PREFIX env variable to
specify different registry values.

Should it be required as part of push/pull mechanics from repos,
it will likely be restored as part of a more general solution
allowing the tagging and publishing based on parameterized
locations vice the hard-coded docker.io

[ch4627]