pg_rewind is now enabled in the default DCS configuration for the crunchy-postgres-ha container.  Therefore, when the default configuration is utilized and Patroni detects that a former primary's timeline has diverged from the current/new primary following a failover, Patroni will attempt to use pg_rewind to synchronize the old primary primary with the new primary, so that it can rejoin the cluster as a replica.
Additionally, the "remove_data_directory_on_rewind_failure"
Patroni configuration setting has been set to "true" in the
default local postgresql configuration for crunchy-postgres-ha
container.  With this setting in place, if a pg_rewind attempt
fails, the PGDATA directory for the PG node will be deleted,
and the replica will be reinitialized (by default using
'pgbackrest restore').

Combined the two settings above ensure that a failed
former primary can successfully rejoin the cluster in the
event that it's timeline has diverged from the current
primary.