Ensure pgAdmin4 displays the actual version being used
Previously the entire config file for pgAdmin4 was being overwritten by
our own local settings. We only need to overwrite a few settings for
this container to work, so the minimum overwritable settings are now
applied.

This also fixes an issue where the directory for logging may not be
present as well, which would lead to a failure at container startup.

Issue: [ch6458]