fixes implicit rule failures for docker builds
The removal of the extraneous docker tag command left some implicit
rules without a recipe. This particular syntax, however, is the
implicit-rule cancel syntax:
http://www.gnu.org/software/make/manual/make.html#Canceling-Rules

To remedy this syntax, make needs to be explicitly told the recipe is
empty via the trailing semicolon.

Other changes:
 * Fixing the implicit rule syntax helped explain my failed attempts to
use a generic rule for upgrades with a single 9.5 exclusion
 * The build rules for non-pg images and upgrade images had the base
dependency moved to them instead of one layer out, to be consistent
with the rest of the build rules