This commit adds monitoring for failed PG starts on replicas, specifically in the form of Kubernetes liveness checks that determine whether or not a replica is in a 'start failed' state, and then reacts accordingly.  Specifically if a liveness probe detects that a replica is in a "start failed" state then the Patroni "reinitialize" REST  endpoint is called on the local Patroni node to reinitialize the replica (please note that the container is never restarted due to a failed liveness check).  This change therefore assists with auto-healing PG clusters, specifically taking action to attempt to bring a replica back online in the event that it is stuck in a "start failed" state.
Additionally, now when a replica is promoted to
primary following a failover (or switchover), a
Patroni 'on_role_change' callback script is now called,
which creates a new pgBackRest backup.  This enables the
best practice of creating a fresh/new backup following a
failover (or switchover), specifically by ensuring a
backup is automatically taken following a failover event.
The creation of the backup following a failover event
(i.e. when a replica is promoted to primary) also has the
added benefit of ensuring that any replicas are able to
initialize or be re-initialized using a recent/up-to-date
backup, preventing issues (e.g. incorrect timelines), in
the event that WAL archives are missing when attempting
to restore from an older backup to bring a replica
online.

Also included are changes to the pgBackRest replica
creation process.  Specifically, a script is now
included to ensure the proper type of restore is
performed.  For instance, if the PGDATA directory is
empty, then a non-delta restore is used.  Further, if
valid PGDATA directory is identified, the a delta
restore is performed.  And lastly, if determined that
the PGDATA directory is invalid (i.e. it is missing the
PGVERSION file), then all files are first deleted
followed by performing a non-delta restore.  These
changes ensure pgBackRest can always be used to
initialize/re-initialize a replica (e.g. even if the
PGDATA directory is invalid), while also preventing
certain pgBackRest warnings from appearing in container
logs (e.g. warning when doing a delta restore into an
empty directory).

Further, readiness probes have been updated for both
the 'crunchy-postgres-ha' and 'crunchyadm' containers.
Specifically, the 'crunchy-postgres-ha' readiness probe
now checks the Patroni "health" endpoint to determine
if the pod is healthy, and is therefore ready to accept
traffic (while also still using the presence of the
'pgha_initialized'file to first determine that the node
was fully initialized).  The 'crunchy-admin' readiness
probe now attempts to connect to the database and
run a query as the 'crunchyadm' user, specifically
connecting via a Unix socket to verify that peer
authentication is working properly for the 'crunchyadm'
user.

And finally, logging has been updated to ensure various
initialization and "start failed" detection activities
are properly logged. Please also note that all changes
discussed above are reflected and implemented within the
'postgres-ha' example for Kube as contained within the
Crunchy Container Suite.