Enable the PostgreSQL HA container to start up as a warm standby
After a recovery is performed, PostgreSQL is in an inconsistent state and may
require some time to replay the write-ahead logs to achieve consistency. While
PostgreSQL is in recovery mode, it is actually inaccessible: a hot standby
would allow connections, but no actions can be taken.

As such, if we know that PostgreSQL is in recovery mode (i.e. having a
recovery.conf file present in PG < 12, and a recovery.signal file present in
PG >= 12), we should start up the PostgreSQL HA instance as a warm standby. This
prevents other services (e.g. Patroni) from making connections, being confused
about PostgreSQL's currently availability, and then doing things that we may
not want to have happen.

Importantly, while PostgreSQL is in recovery mode, it allows us to indicate that
the container is not ready, which can have ramifications in a Kubernetes
environment.

This also fixes a bug where we would drop a file used as part of our readiness
checks before PostgreSQL is ready, as we constitute "ready" as being in a
consistent state.

Issue: [ch6609]