This commit ensures the pgBackRest settings required to enable both local and AWS S3 storage are merged with the existing pgBackRest configuration settings during cluster initialization, instead of overwriting the existing pgBackRest configuration settings.
Additionally, when an existing database is started
manually in order to then be converted to a
Patroni cluster, the container will now wait
indefinitely for the cluster to reach a consistent
state and become ready.  Additionally, if the
container detects that PostgreSQL is no longer
running while waiting for the DB to reach a
consistent state, the container will attempt to
start the database again (e.g. in the event that
recovery failed due to unavailable WAL on the first
start attempt due to a connectivity issue).  This
will ensure the DB will keep attempting to recover
without the container crashing, providing the
opportunity to auto-recover and/or intervene and
address any issues that may exist.