This commit adds a few changes to the pgBackRest replica initialization process.  First, if an empty PGDATA directory is found when initializing a replica, a command is executed to created the empty PGDATA directory if it does not already exist.  This is to ensure the directory exists in the event that it was deleted, and node is now re-initializing as a replica.
Also, when an existing database is being started manually
prior to initializing Patroni, the startup script determines
if the postgresql.conf file needs to be cleaned.  Specifically,
if the Patroni bootstrap configuration file is configured to
use a custom PG config file using the custom_config
parameter, or if a postgresql.base.conf file is present in the
PGDATA directory, then the script assume those files are the
base postgresql.conf parameters for the database in
accordance with the Patroni documentation for applying PG
configuration to a cluster, and therefore cleans out the
contents of the existing postgresql.conf file if it exists
(otherwise an empty file will simply be created).  This ensures
that the settings from previous Patroni clusters that may no
longer be valid (e.g. invalid dir names) do not prevent the DB
from starting successfully.  Once Patroni is initialized, the
postgresql.conf will be populated with the dynamic
configuration defined for the cluster.

If a custom or base config file is not found, then the
postgresql.conf will remain untouched, and will then become
the base configuration is accordance with the Patroni
documentation for apply PG configuration.