With this change, it is possible to apply the required SSL configuration to enable SSL in the PG database. Specifically, by mounting the appropriate certificates in the pgconf directory, it is possible to  configure the 'ssl_key_file', 'ssl_cert_file', 'ssl_ca_file' and 'ssl_crl_file' postgresql.conf settings for the cluster via the DCS, effectively enabling the use of SSL and client certificates when connecting to and authenticating into the database.
In support of configuring the DCS for SSL, the
mechanism for detecting that the local node has been
initialized has been updated.  Now, in addition to
verifying that the node is in a running state, a check
is performed to first determine whether the local node
has the primary or replica role, and then whether or not
the primary or replica endpoint (depending on the role
identified) is returning status code 200.  This ensures
that all initial configuration has been applied to the
node prior to updating the DCS with the SSL configuration
using a JSON patch.

And finally, it is now possible to specify the certificates
needed to allow the replication user to authenticate via
client certificate in lieu of using a replication password.