Add support for PostgreSQL tablespaces in the HA containers
Tablespaces can be used to spread out PostgreSQL workloads across
multiple volumes, which can be used for a variety of use cases:

- Partitioning larger data sets
- Putting data onto archival systems
- Utilizing hardware (or a storage class) for a particular database
object, e.g. an index

and more.

When the appropriate environmental variables that govern the tablespaces
are passed into the crunchy-postgres-ha and crunchy-postgres-gis HA
containers, the directories where the tablespaces reside are initialized
so PostgreSQL can appropriately access them. Additionally, when a
PostgreSQL cluster is initialized, the tablespace object is
automatically created (via CREATE TABLESPACE) and the initial user is
granted access to it.

Issue: [ch6495]