Reorganize crunchy-postgres-ha Container bin Directory
This commit reorganizes the various bash and SQL scripts under the
$CCPROOT/bin/postgres-ha directory into sub-directories based on the
specific functionality included in each script.  This is to better
organize these scripts and add clarity around the various scripts
packaged with the crunchy-postgres-ha (and GIS equivalent) containers.

The following list defines the various sub-directories added to the bin
directory, along with a description of the specific scripts included in
each new directory:

- bootstrap: cluster bootstrap & initialization scripts
- callbacks: patroni callback scripts
- common: common functions utilized across various scripts
- health: health check scripts, e.g. for readiness & liveness probes
- modules: module initialization scripts, e.g. for pgmonitor & pgaudit
- pgbackrest: pgbackrest configuration and initialization scripts
- sql: sql scripts to be executed during bootstrap & initialization

The bootstrap-postgres-ha.sh script remains at the root of the bin
directory as the entrypoint for the container.  Additionally, the
setup.sql script in the bin directory for the posgres-ha-gis container
has been moved to a 'sql' sub-directory in order to remain consistent
with this new directory structure.