Create pgbouncer helpers functions as part of container initialization
The PostgreSQL Operator would provide some embedded SQL to help with
this process, but arguably was doing too much work. The requirements
it has for using pgBouncer are fairly straightforward:

- Have a pgBouncer user
- Have a pgBouncer schema created in all accessible databases
- Have a "get_auth" function that follows the "auth_query" pattern
defined in the pgBouncer docs[1]
- Lock everything down as much as possible regarding what the pgbouncer
user can do

This commit provides the scripts to initialize the pgbouncer user
and machinery in the PostgreSQL HA container, but defers to the
container user to determine if they get executed.

[1] http://www.pgbouncer.org/config.html#auth_query