Updated Directory for UID & pgBackRest Info Scripts
Updated the directory for the uid_postgres_no_exec.sh script where
it is called in the bootstrap script to align with the new 'bin'
directory structure implemented in the crunchy-postgres-ha
container.

Also corrected the location of the script utilized to load
pgBackRest environment variables in the pgbackrest_info.sh
script that is utilized by pgMonitor.