Support for Standby Clusters
This commit adds support for standby clusters, i.e. clusters
replicating from another external database using the WAL generated
from that database. In order to support standby cluster creation, a
new environment variable has been created named PGHA_STANDBY.  When
this environment variable has been set to "true", the configuration
defined in a new config file called postgres-ha-standby.yaml is
merged into the bootstrap YAML configuration file generated to create
the cluster.  This provides the configuration needed replicate from
another PostgreSQL database outside of the cluster using the WAL
generated by that cluster.  This is specifically accomplished using
pgBackRest, which leverages the pgBackRest repository of the external
database to bootstrap the cluster, and then continuously obtain any
WAL generated.

Additionally, this commit also establishes a mechanism for
determining the type of repo that should be utilized for generating
replicas, e.g. if both AWS S3 and local in-cluster repositories
exist.  This is accomplished by checking for the presence of a file
called 'replica-bootstrap-repo-type', which is responsible for
being populated with a valid repo type per the pgBackRest
documentation (or it can be omitted or left blank to apply the
default repo type).  This strategy allows the replica creation
source for a replica to be modified without requiring a restart of
the container (e.g. in the event the replica needs to be
reinitialized using a different repo type than used during its initial
creation).