Allow for pgBackRest Repo Type Override
This commit allows the pgBackRest repo type (set using the
PGBACKREST_REPO_TYPE environment variable within the
crunchy-postgres-ha container) to be overridden.  This is
needed to ensure the pgBackRest repo type can be contextually set to
a value other than the value set for PGBACKREST_REPO_TYPE within the
environment during container inititialization without requiring a
container restart to modify the env var.  For instance, a standby
cluster will want to ensure repo type "s3" is always selected as the
repo type while the cluster is in standby mode, since the local repo
(if enabled) will be empty, requiring the use of "s3" to bootstrap the
cluster.

In support of this capability , a new script has been created that is
now utilized across various scripts to load the proper configuration
for various pgBackRest commands (pgbackrest-set-env.sh). This script
leverages the pgbackrest_env.sh script created during container start
up to load the base pgBackRest configuration, while then providing the
ability to dynamically override any configuration settings as needed
for the specific command being executed (e.g. the repo type, as
described above).

In the case of overriding the repo type, the script checks for the
presence of a 'replica-bootstrap-repo-type' configuration file that
explicitly defines the type of repo that should be utilized when
running a pgBackRest command.  If this config file is present and
contains a value, then that value is utilized as the value for
PGBACKREST_REPO_TYPE when running a pgBackRest command.  Otherwise, the
value already present in the environment for PGBACKREST_REPO_TYPE as
set when the container was initialized is utilized instead.