Move pgbackrest-set-env.sh Into Proper Directory
This commit moves the pgbackrest-set-env.sh script into the
'pgbackrest' folder underneath the 'bin' directory for the
crunchy-postgres-ha container.  This ensures the location of this
script aligns with the new directory structure established for
storing the various bash scripts utilized by the crunchy-postgres-ha
container.  Additionally, all references to the pgbackrest-set-env.sh
script have been updated to specify the proper location for the script
(i.e. directory /opt/cpm/bin/pgbackrest).