Allow regular PostgreSQL user to create objects in its own schema
The "setup.sql" script that runs several bootstrap commands for a
PostgreSQL database that is deployed by the Operator includes adding
a user schema (a standard part of the search patch[1], but makes the
mistake of not allowing the user to access the schema. This patch
adds the "AUTHORIZATION" directive to the "CREATE SCHEMA" command to
ensure that the user can actual utilize the schema, making the feature
actually useful.

Issue: CrunchyData/postgres-operator#1334

[1] https://www.postgresql.org/docs/current/ddl-schemas.html#DDL-SCHEMAS-PATH