Refresh init_role During PG HA Init
The loop used to determine if PostgreSQL is running during
initialization of the PG HA container now refreshes the role
variable on each pass of the loop as used to determine if it
is a primary or replica.  This is needed in situations where
the node might initially identify as a replica, but then is
immediately promoted to master (e.g. when a cluster is
started after standby mode has been enabled).