Only Run Post Init if PGHA_INIT is 'true'
This commit updates the initialization monitor for the postgres-ha
container to only run initialization logic (e.g. creating the
crunchyadm user, applying SSL configuration, etc.) if the PGHA_INIT env
var is 'true'.  Previously, the initialization monitor would attempt to
detect the role of the local node, and then use the specific "role"
endpoint (the "replica", "master", etc. endpoints) to determine if the
local node is running.  And if the node was identified as a "master",
then the initialization logic would be run.  However, this can be
problematic if the role initially starts as a certain role, e.g. a
"replica"), but is immediately promoted to "master".  Specifically,
this could result in a node running initialization logic, even though
it started as a replica (and that logic therefore shouldn't be run_).

Being that the Operator ensures that the PGHA_INIT flag is only "true"
when needed, e.g. when initializing a cluster for the first time, and
then immediately sets it to "false", this env var provides an effective
solution for triggering the initialization logic only when first
creating a cluster, and then ensuring it is then never run again unless
needed.