Allow recovery_target_action to be 'pause'
When hot_standby is disabled and recovery_target_action is unset or set
to 'pause', PostgreSQL >= 9.5 will shut down at the end of recovery. To
allow 'pause' and still detect successful recovery, we must connect and
call pg_is_in_recovery().

We cannot delegate this last step of point-in-time recovery to Patroni
because it interprets a data directory restored by pgBackRest to be a
PostgreSQL crash. To remedy the suspected crash, Patroni clears all
recovery settings before starting PostgreSQL.

Issue: [ch7621]