Reduce Use of Manual Database Starts
With this change, the database will only be manually started under the
following conditions:
- A PITR is detected
- A non-Patroni standalone database is being converted to a Patroni
cluster
Otherwise, in all other scenarios Patroni start the database, handling
any recovery procedures as needed.

To auto-detect whether or a PITR is being performed, the bootstrap
script for the HA container now checks to see if a 'recovery.signal'
(PG v12 and above) or a 'recovery.conf' (PG v11 and below) is
present, along with whether not any 'recovery_target' settings
are configured for the restore.  To detect if a non-Patroni standalone
database is being converted to a Patroni cluster, the bootstrap
script checks to see if there is an existing PGDATA directory which
does not contain a patroni.dynamic.json.  If this file is present, an
existing Patroni cluster is assumed and Patroni will handle the
recovery.

Additionally, 'post-existing-init.sql' is now only run when converting
a standalone database to a Patroni cluster, and the
'pgha-on-role-change.sh' now exits immediately (and gracefully) if
the cluster has not yet been initialized.  A check is also now in place
to see if the 'crunchyadm' user already exists prior to attempting
to create it.  And finally, all enhancement modules (pgMonitor, pgAudit
Analyze, etc.) are now always run (if enabled) when initializing a
primary, regardless of whether that primary is being initialized as
part of a brand new database, or if it is being initialized as part of
a clone or restore workflow.