Update crunchyadm Logic
The crunchyadm.sh script has been updated to use a variable in the psql
command for the username representing the user being created.  This
includes switching to the '-f' instead of '-c' as needed to support
variables.  Also removed the use of 'grep' as it was not needed (i.e.
the output of the 'psql' command can simply be inspected instead).