Update configuration files
This commit makes a pass at the configuration files to update them and
have them leverage the PostgreSQL defaults more, given most
customizations in the files are doing exactly that as is. This also
updated terminology to use language that is accepted and promoted by
the PostgreSQL community for describing distributed systems.