Enable replication connections to use certificate-based authentication
Certificate-based authentication[1] is a powerful PostgreSQL feature
that allows for a PostgreSQL client to authenticate using a TLS
certificate. While there are a variety of permutations for this can be
set up, we can at least create a standardized way for enabling the
replication connection to authenticate with a certificate, as we do have
a known certificate authority.

This introduces support for certificate-based authentication for the
replication user, presuming that the CN on the certificate is that of
`PATRONI_REPLICATION_USERNAME` (default is "primaryuser"). If the CN
on the certificate is different, one must include a "pg_ident"
customization as part of their configuration.

If TLS is enabled on the container and a Secret that contains the
replication user's TLS credentials is mounted to the container, then
certificate-based authentication for this user is automatically set up.
As the CA must be appropriately mounted for this to work, we are able
to set PGSSLMODE to use "verify-ca" by default, to include a stronger
guarnatee of validity between the instances (though not as strong as
"verify-full", but we cannot make that level of assumption about the
environment).

Issue: [ch8387]

[1] https://www.postgresql.org/docs/current/auth-cert.html