Init Job Support
This commit adds support to allow the 'crunchy-postgres-ha' container
to be run as an initialization Job.  This can be used to initialize
a new PostgreSQL database using an existing data source (e.g. a
pgBackRest backup), and then cleanly shut the database down once
initialization is complete (e.g. a pgBackRest restore has finished
successfully and recovery has completed) so that it can then be
cleanly started again when needed.  This specifically allows the
PostgreSQL Operator to initially run batch Job to populate the PGDATA
volume for a new cluster, and then subsequently mount and use
that same volume to bring a new PostgreSQL cluster online.

In support of running the container as an initialization Job, a new
custom bootstrap method called 'pgbackrest_init' has been created.
When this bootstrap method is configured, a 'pgbackrest restore' is
performed when initializing a new PostgreSQL cluster (instead of
running 'initdb' to initialize from scratch).  A 'RESTORE_OPTS'
environment variable is also provided that allows for further
customization of the 'pgbackrest restore' command (e.g.
specifying a '--type' and 'target' to use a PITR to initialize
the cluster).

Also, the 'post-existing-init.sql' script is now executed from the
post-restore script following the following the initialization of
a cluster using the 'pgbackrest_init' bootstrap method.  This
creates and/or updates the credentials for the standard user
created with all new PostgreSQL clusters, and can be further
customized to run additional SQL upon bootstrap completion when
using this bootstrap method.