Ensure pgBackRest Settings Work When Local & S3 Storage is Enabled
This update is the companion update to the Postgres Operator PR which
ensures that the 's3-verify-tls' & 's3-uri-style' settings work as expected
when both local & S3 storage are enabled for pgBackRest when creating a
pgcluster. Currently, when a new cluster is created using Local and AWS S3
storage (e.g. by specifiying --pgbackrest-storage-type=local,s3 with the
pgo create command), stanza creation fails and the cluster is not initialized.

Since this is failing because the 'non S3' pgBackRest command runs when
the repo1-s3-verify-tls flag is detected in the environment. This fix updates
the pgBackRest flag implementation so that the repo1-s3-verify-tls setting
is applied only for the second pgbackrest command only, i.e. the command using
--repo-type-s3, and not apply it when running the first pgbackrest command for
the local repository. This is done by using the --no-repo1-s3-verify-tls
command flag rather than the relevant environment variable or the
related configuration file setting. This method allows both commands to
execute as expected.