Only Run Post-Init for Non-Standby Clusters
With this commit the post-initialization logic that is typically run
after the primary initializes within a new PostgreSQL cluster (e.g.
waiting for recovery to complete and adding enhancement modules) is
no longer run when bootstrapping a standby cluster.  This is
specifically because we expect a standby cluster to remain in recovery
(and therefore remain in a read-only mode), and should obtain
anything written to the database from the Primary cluster it is
replicating from.