Remove crunchy-grafana & crunchy-prometheus
This commit fully removes the 'crunchy-grafana' and
'crunchy-prometheus' containers from the Crunchy Container Suite.  This
includes the full removal of all 'crunchy-grafana' and
'crunchy-prometheus' Dockerfiles, examples, build artifacts,
documentation and references.

This change specifically supports the full migration of the metrics
infrastructure to the PostgreSQL Operator, with the PostgreSQL Operator
Metrics infrastructure now leveraging the upstream Grafana and Prometheus
containers instead of the 'crunchy-grafana' and
'crunchy-prometheus' containers.