Add Bootstrap Method for Existing PGDATA Dir
This commit adds a new Patroni bootstrap method called 'existing_init'
that initializes a new PostgreSQL cluster using an existing PGDATA
directory.  Now when the PGHA_INIT flag is 'true' and the
crunchy-postgres-ha container detects that the PGDATA directory for the
cluster is not empty, the 'existing_init' bootstrap method is executed.

This bootstrap method prevents Patroni from treating the initial start
of the database as the recovery of a former cluster (as it typically
would when the PGDATA directory isn't empty), but rather as the start
of a brand new cluster.  This means the post-initialization script is
executed, the instance is started cleanly as a primary, and any new
configuration settings specified when creating the cluster are applied
(instead of restoring the configuration from a previous cluster).

Additionally, various PITR restore and manual DB start logic has been
removed since it is no longer needed (Patroni now handles all DB starts
and PITR's).

Issue: [ch8719]
Issue: [ch8721]