Proper File Permissions for Existing PGDATA Dir
When bootstrapping a new PostgreSQL cluster using an
existing/pre-populated PGDATA directory, the directory file permissions
are now explicitly set to 0700.  This is to ensure any additional file
permissions added to the volume (e.g. when Kubernetes applies fsGroups)
are removed as needed to ensure the database can be properly started.

Issue: [ch9156]