Add wal2json to the PostgreSQL containers
wal2json is great for applications that want to leverage
logical decoding in an output format that is easy to understand,
let alone parse. The committer of the patch has even used it
himself in the past. So, let's bring it into the containers.

Issue: [ch9125]
Issue: CrunchyData/postgres-operator#1816
Issue: CrunchyData/postgres-operator#1841